+++ Firmware/AngleMod/pic16f152-stubs/src/pic16f152.cpp
@@ -40,7 +40,7 @@ volatile struct PIE0bits PIE0bits;
 volatile struct INTCONbits INTCONbits;
 volatile struct NVMCON1bits NVMCON1bits;
 
-volatile uint8_t NVMADR;
+volatile uint16_t NVMADR;
 volatile uint8_t NVMCON2;
 volatile uint8_t NVMDATL;
